requirements.txt
pyautogen==0.2.25
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
openai==1.35.3
gunicorn==22.0.0 